Moved api.py to /ctf/__init__.py
